near_real_time/requirements.txt
kafka-python==2.0.2
lz4==4.3.2
pydantic~=1.10.5
pandas~=1.5.3
numpy==1.23.5
scikit-learn==1.2.1
python-dotenv==1.0.0